core: rework usage of ReportReconcileResult
ReportReconcileResult should never be given an object that is nil. It is
impossible to force compile-time checking for this because the object's
type is an interface. We can't force this, but we can rework the
function definition to handle more cases where the object may not be
returned fully-complete, and we can rework callers to return structs
(not pointers-to-structs) so it is less likely a caller will pass nil
to the function.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>
(cherry picked from commit c92c6fbc36fbfd01f9f90aca08e4f392b0061f10)
(cherry picked from commit ae1eab91fc9990a72aa347ce0fa68b317a32a7da)